バグ改修50%程度完了

+++ Project/src/ec/UserDeleteResult.java
@@ -9,6 +9,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import beans.UserDataBeans;
 import dao.UserDAO;
 
 @WebServlet("/UserDeleteResult")
@@ -23,9 +24,26 @@ public class UserDeleteResult extends HttpServlet {
 			String id_str = request.getParameter("id");
 			int id = Integer.parseInt(id_str);
 
+			// idの一致するユーザまたは管理者としてログインセッションがない場合、ログイン画面にリダイレクトさせる
+			UserDataBeans loginUdb =(UserDataBeans)session.getAttribute("userInfo");
+			if(loginUdb == null || !(loginUdb.getLoginId().equals("admin") || loginUdb.getId() == id) ) {
+				// セッションにリターンページ情報を書き込む
+				session.setAttribute("returnStrUrl", "UserDeleteResult");
+				// ログイン画面にリダイレクト
+				response.sendRedirect("Login");
+				return;
+			}
+
 			//idを引数にして、idに紐づくユーザ情報を削除する
 			UserDAO userDao = new UserDAO();
 			userDao.deleteUser(id);
+
+			// 管理者としてログイン	してた場合
+			if(loginUdb.getLoginId().equals("admin")) {
+				// ユーザ一覧のサーブレットにリダイレクト
+				response.sendRedirect("UserList");
+				return;
+			}
 
 			// ログイン時に保存したセッション内のユーザ情報を削除
 			session.setAttribute("isLogin", false);
